pocs/scikit-surprise-rec/requirements.txt
numpy
transformers
sentence-transformers
seaborn
torch
torchvision
matplotlib
pandas
scikit-learn
nltk
gensim
tensorflow
keras
opencv-python
fastapi
uvicorn
scikit-surprise
